+++ Src/XmlDocInspections.Plugin/Settings/XmlDocInspectionsSettings.cs
@@ -1,4 +1,5 @@
-﻿using JetBrains.Application.Settings;
+﻿using System.Diagnostics.CodeAnalysis;
+using JetBrains.Application.Settings;
 using JetBrains.ReSharper.Resources.Settings;
 
 namespace XmlDocInspections.Plugin.Settings
@@ -7,24 +8,30 @@ namespace XmlDocInspections.Plugin.Settings
     /// Settings class for Xml Doc inspections.
     /// </summary>
     [SettingsKey(typeof(CodeInspectionSettings), "Xml Doc Inspections")]
+    [SuppressMessage("ReSharper", "NotNullMemberIsNotInitialized")]
     public class XmlDocInspectionsSettings
     {
         private const AccessibilitySettingFlags DefaultAccessibilityFlags =
             AccessibilitySettingFlags.Public | AccessibilitySettingFlags.Protected | AccessibilitySettingFlags.ProtectedOrInternal;
 
+
         [SettingsEntry(DefaultAccessibilityFlags, "Type accessibility")]
-        public AccessibilitySettingFlags TypeAccessibility;
+        public readonly AccessibilitySettingFlags TypeAccessibility;
 
         [SettingsEntry(DefaultAccessibilityFlags, "Type member accessibility")]
-        public AccessibilitySettingFlags TypeMemberAccessibility;
+        public readonly AccessibilitySettingFlags TypeMemberAccessibility;
+
+        [SettingsEntry("JetBrains.Annotations.PublicAPIAttribute", "Include types/members with attributes (comma separated full names)")]
+        public readonly string IncludeAttributeFullNames;
+
 
         [SettingsEntry(false, "Exclude constructors")]
-        public bool ExcludeConstructors;
+        public readonly bool ExcludeConstructors;
 
         [SettingsEntry(false, "Exclude members which override super/base members")]
-        public bool ExcludeMembersOverridingSuperMember;
+        public readonly bool ExcludeMembersOverridingSuperMember;
 
         [SettingsEntry("Tests$", "Project exclusion regex")]
-        public string ProjectExclusionRegex;
+        public readonly string ProjectExclusionRegex;
     }
 }